requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-bootstrap4==0.0.8
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==2.8
oauthlib==3.0.1
Pillow==5.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
sqlparse==0.3.0
urllib3==1.25.3
